backend/requirements.txt
Django>=6.0.2
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
python-docx>=1.2.0
pymupdf>=1.27.2
google-genai>=1.66.0
python-dotenv>=1.2.1
lxml>=6.0.2
whitenoise>=6.6.0
axios
